requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
ipython==3.1.0
numpy==1.9.2
pyzmq==14.6.0
scikit-learn==0.16.1
scipy==0.15.1
speech-features==0.1
tornado==4.2
wsgiref==0.1.2
